Implementation of issue #21
Completely rewrote how instructions are created.  Removed functionality from PushP and moved it to Plush.  The PushP implementation was causing the code to crash.

+++ SOS/PushP/CodeOps.h
@@ -102,7 +102,8 @@ namespace Push
 		return 1; 
 	}
 
-	inline unsigned _append() {
+	inline size_t _append()
+	{
 		if (1 + first<Code>()->size() + second<Code>()->size() >= env.parameters.max_points_in_program) 
 			return 1;
